requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.2.0
python3-discogs-client>=2.7
